changes for recompile only with admin access
Former-commit-id: d6569b247a1e1ad1dd812989b73b8dc0f02394b2

+++ Assemblies/ProteanCMS/Tools/xmlTools.xsltExtensions.vb
@@ -2034,27 +2034,36 @@ Partial Public Module xmlTools
                 Else
                     'Dim fsh As New Protean.fsHelper(myWeb.moCtx)
                     If Not myWeb.moRequest("rebundle") Is Nothing Then
-                        bReset = True
+                        ''code for deleting script.js file from the bundle folders.
+                        Dim sSessionId As String = String.Empty
+                        Dim nDuration As Integer = 5
+                        Dim bIsAdmin As Boolean = False
+                        If (myWeb.moRequest("SessionId") IsNot Nothing) Then
+                            sSessionId = myWeb.moRequest("SessionId")
+                            bIsAdmin = myWeb.RestoreRedirectSession(sSessionId, nDuration, True)
+                        End If
 
-                        'code for deleting script.js file from the bundle folders.
-                        Dim pathSplitString As String() = TargetPath.Split(New Char() {"/"c})
 
-                        Dim paths As String = myWeb.goServer.MapPath("/" & myWeb.moConfig("ProjectPath") & "js/" & pathSplitString(1))
-                        Dim rootfolder As New DirectoryInfo(myWeb.goServer.MapPath("/" & myWeb.moConfig("ProjectPath") & "js/" & pathSplitString(1)))
-                        If rootfolder.Exists Then
-                            For Each filepath As String In Directory.GetFiles(paths)
-                                File.Delete(filepath)
-                            Next
-                            'Delete all child Directories
-                            For Each dir As String In Directory.GetDirectories(paths)
-                                For Each filepath As String In Directory.GetFiles(dir)
+                        If myWeb.mbAdminMode Then
+                            bReset = True
+                            Dim pathSplitString As String() = TargetPath.Split(New Char() {"/"c})
+                            Dim paths As String = myWeb.goServer.MapPath("/" & myWeb.moConfig("ProjectPath") & "js/" & pathSplitString(1))
+                            Dim rootfolder As New DirectoryInfo(myWeb.goServer.MapPath("/" & myWeb.moConfig("ProjectPath") & "js/" & pathSplitString(1)))
+                            If rootfolder.Exists Then
+                                For Each filepath As String In Directory.GetFiles(paths)
                                     File.Delete(filepath)
                                 Next
-                            Next
+                                'Delete all child Directories
+                                For Each dir As String In Directory.GetDirectories(paths)
+                                    For Each filepath As String In Directory.GetFiles(dir)
+                                        File.Delete(filepath)
+                                    Next
+                                Next
+                            End If
+                            myWeb.moCtx.Application.Remove(TargetPath)
                         End If
-                        myWeb.moCtx.Application.Remove(TargetPath)
                     End If
-                    Dim bAppVarExists As Boolean = False
+                        Dim bAppVarExists As Boolean = False
                     If Not myWeb.moCtx.Application.Get(TargetPath) Is Nothing Then
                         bAppVarExists = True
                     End If
